refactor: make package name conventional
