+++ cmd/http_server/main.go
@@ -3,7 +3,7 @@ package main
 import "github.com/team-inu/inu-backyard/infrastructure/fiber"
 
 func main() {
-	fiberServer := fiber_handler.NewFiberServer()
+	fiberServer := fiber.NewFiberServer()
 
 	fiberServer.Run()
 }
