+++ infrastructure/fiber/controller/student_controller.go
@@ -1,4 +1,4 @@
-package fiber_controller
+package controller
 
 import (
 	"github.com/gofiber/fiber/v2"
